+++ libclang-bindings/src/Clang/LowLevel/Core.hs
@@ -217,6 +217,7 @@ module Clang.LowLevel.Core (
   , clang_Location_isFromMainFile
     -- * File manipulation routines
   , clang_getFileName
+  , clang_File_tryGetRealPathName
     -- * Debugging
   , clang_breakpoint
     -- * Exceptions
@@ -2008,6 +2009,16 @@ clang_Location_isFromMainFile location = liftIO $
 -- <https://clang.llvm.org/doxygen/group__CINDEX__FILES.html#ga626ff6335ab1e0a2b8c8823301225690>
 clang_getFileName :: MonadIO m => CXFile -> m Text
 clang_getFileName file = liftIO $ preallocate_$ wrap_getFileName file
+
+-- | Get the real, canonical path name of the given file.
+--
+-- Returns the empty string for virtual files (files that are not saved to
+-- disk, such as unsaved files passed to @clang_parseTranslationUnit@).
+--
+-- <https://clang.llvm.org/doxygen/group__CINDEX__FILES.html#ga2becf44b7e67e25038e6e4fdb3e22a96>
+clang_File_tryGetRealPathName :: MonadIO m => CXFile -> m Text
+clang_File_tryGetRealPathName file =
+    liftIO $ preallocate_$ wrap_File_tryGetRealPathName file
 
 -- | Retrieve the contents of the given file that is loaded in the given
 -- translation unit.